Avoid failing dispose when directory is already deleted

+++ src/Arkivverket.Arkade.Core.Tests/UnitTestUtilities/DisposableDirectory.cs
@@ -28,6 +28,7 @@ public class DisposableDirectory : IDisposable
 
     public void Dispose()
     {
-        _directory.Delete(true);
+        if (_directory.Exists)
+            _directory.Delete(true);
     }
 }
